Fix problems with EndermanMixin mappings on production servers (#21)
Explicitly target the remapped function name for the canStart method
There appears to be a bug in mixin that causes the canStart method to not get remapped properly in production environments. A simple fix for now is to just target both the clean and the remapped function name so the code will work both in development and production environments.

Reviewed-on: https://hogwarts.bits.team/git/Bits/BitsVanilla/pulls/21
Reviewed-by: Kallum <kallummj071001@gmail.com>